Refactoring

+++ src/arch/i686/driver/cmos.c
@@ -4,6 +4,8 @@
  * @copyright MIT
  */
 
+#include <stddef.h>
+
 #include "io.h"
 #include "cmos.h"
 #include "idt.h"
@@ -12,11 +14,11 @@ CMOS cmos = {
     .values = {},
 };
 
-int bcd_to_decimal(const int bcd)
+int32_t bcd_to_decimal(const int32_t bcd)
 {
-    const int upper_nibble = (bcd & 0xF0) >> 4;
-    const int lower_nibble = bcd & 0x0F;
-    const int decimal = (upper_nibble * 10) + lower_nibble;
+    const int32_t upper_nibble = (bcd & 0xF0) >> 4;
+    const int32_t lower_nibble = bcd & 0x0F;
+    const int32_t decimal = (upper_nibble * 10) + lower_nibble;
     return decimal;
 };
 
@@ -29,7 +31,7 @@ void cmos_dump(CMOS *self)
         asm_outb(CMOS_PORT_INDEX, cmos_register);
         const uint8_t value = asm_inb(CMOS_PORT_DATA);
 
-        for (int i = 0; i < 100; i++)
+        for (size_t i = 0; i < 100; i++)
         {
             asm_do_nop();
         };
@@ -41,16 +43,16 @@ void cmos_dump(CMOS *self)
 Date cmos_date(CMOS *self)
 {
     cmos_dump(self);
-    const int bcd_weekday = self->values[6];
-    const int bcd_day = self->values[7];
-    const int bcd_month = self->values[8];
-    const int bcd_century = self->values[32];
-    const int bcd_year = self->values[9];
+    const int32_t bcd_weekday = self->values[6];
+    const int32_t bcd_day = self->values[7];
+    const int32_t bcd_month = self->values[8];
+    const int32_t bcd_century = self->values[32];
+    const int32_t bcd_year = self->values[9];
 
-    const int weekday = bcd_to_decimal(bcd_weekday);
-    const int day = bcd_to_decimal(bcd_day);
-    const int month = bcd_to_decimal(bcd_month);
-    const int year = bcd_century * 10 + bcd_to_decimal(bcd_year);
+    const int32_t weekday = bcd_to_decimal(bcd_weekday);
+    const int32_t day = bcd_to_decimal(bcd_day);
+    const int32_t month = bcd_to_decimal(bcd_month);
+    const int32_t year = bcd_century * 10 + bcd_to_decimal(bcd_year);
 
     Date date = {
         .weekday = weekday,
